+++ lib/api/api_service.dart
@@ -5,6 +5,7 @@ import 'package:http/http.dart' as http;
 
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+import 'package:x_qrcode/attendees/checkin_exception.dart';
 
 import 'package:x_qrcode/event/events_screen.dart';
 import 'package:x_qrcode/organization/model/user_model.dart';
@@ -42,6 +43,30 @@ class ApiService {
     }
   }
 
+  Future<List<Attendee>> getAttendees() async {
+    final user =
+        User.fromJson(jsonDecode(await storage.read(key: STORAGE_KEY_USER)));
+    final accessToken = await storage.read(key: STORAGE_KEY_ACCESS_TOKEN);
+    final event =
+        Event.fromJson(jsonDecode(await storage.read(key: STORAGE_KEY_EVENT)));
+
+    final response = await http.get(
+        '${DotEnv().env[ENV_KEY_API_URL]}/${user.tenant}/events/${event.id}/attendees',
+        headers: {HttpHeaders.authorizationHeader: "Bearer $accessToken"});
+
+    if (response.statusCode == 200) {
+      final rawAttendees = jsonDecode(response.body);
+      final attendees = List<Attendee>();
+      for (var rawEvent in rawAttendees) {
+        attendees.add(Attendee.fromJson(rawEvent));
+      }
+      attendees.sort((a, b) => a.firstName.compareTo(b.firstName));
+      return attendees;
+    } else {
+      throw Exception('Failed to load attendees');
+    }
+  }
+
   Future<List<Attendee>> getVisitors() async {
     final user =
         User.fromJson(jsonDecode(await storage.read(key: STORAGE_KEY_USER)));
@@ -54,15 +79,35 @@ class ApiService {
         headers: {HttpHeaders.authorizationHeader: "Bearer $accessToken"});
 
     if (response.statusCode == 200) {
-      final _rawVisitors = jsonDecode(response.body);
-      final _visitors = List<Attendee>();
-      for (var rawEvent in _rawVisitors) {
-        _visitors.add(Attendee.fromJson(rawEvent));
+      final rawVisitors = jsonDecode(response.body);
+      final visitors = List<Attendee>();
+      for (var rawEvent in rawVisitors) {
+        visitors.add(Attendee.fromJson(rawEvent));
       }
-      _visitors.sort((a, b) => a.firstName.compareTo(b.firstName));
-      return _visitors;
+      visitors.sort((a, b) => a.firstName.compareTo(b.firstName));
+      return visitors;
     } else {
       throw Exception('Failed to load visitors');
+    }
+  }
+
+  Future<bool> toggleCheck(String id, bool check) async {
+    final user =
+        User.fromJson(jsonDecode(await storage.read(key: STORAGE_KEY_USER)));
+    final accessToken = await storage.read(key: STORAGE_KEY_ACCESS_TOKEN);
+    final event =
+        Event.fromJson(jsonDecode(await storage.read(key: STORAGE_KEY_EVENT)));
+
+    final response = await http.patch(
+        '${DotEnv().env[ENV_KEY_API_URL]}/${user.tenant}/events/${event.id}/attendees/$id',
+        headers: {HttpHeaders.authorizationHeader: "Bearer $accessToken"},
+        body: {'checked': '$check'});
+
+    if (response.statusCode == 200) {
+      return check;
+    } else {
+      throw CheckInException(
+          'Cannot check-in attendee, status: ${response.statusCode}, message: ${response.body}');
     }
   }
 }
